Minidata, infant growth of 0-2 years, three children

Longitudinal height, weight, head circumference measurements during ages 0-2 years for a representative sample of 3 Dutch children born in 1988-1989, including covariates. For demo purposes.

Format

A list with two components:

child

Tibble with 3 rows and 14 columns containing child-level covariates

time

Tibble with 30 rows and 78 columns containing growth data and developmental milestones per time point.

The child tibble contains the following variables:

id

ID, unique id of each child (numeric)

src

Source, here "smocc" (character)

dob

Data of birth "smocc" (character)

sex

Sex, "male" or "female" (character)

etn

Etnicity, "MA", "NL", or "TU" (character)

edu

Educational level (typically mother), "low", "middle", "high" (character)

#'
ga

Gestational age in completed, weeks (numeric)

bw

Birth weight in grammes (numeric)

twin

Twin, 0 = no, 1 = yes (numeric)

smo

Mother smoked during pregnancy, 0 = no, 1 = yes (numeric)

agem

Mother age when giving birth (years) (numeric)

hgtm

Height of mother in cm (numeric) (numeric)

hgtf

Height of father in cm (numeric) (numeric)

name

Child nickname

The time tibble contains the following variables:

src

Source, here "smocc"

id

ID, unique id of each child

rec

Record number

nrec

Number of child records, 6+

dob

Data of birth (character)

dob

Data of measurement (character)

age

Decimal age

sex

Sex, "male" or "female"

etn

Etnicity, "MA", "NL", or "TU"

ga

Gestational age in completed weeks

bw

Birth weight in grammes

hgt

Height measurement in cm

wgt

Weight measurement in kg

hdc

Head circ measurement in cm

dsc

D-score, D-unit

bds879

DDI milestone, BDS number 879

bds...

DDI milestone, BDS number ...

Source

Herngreen WP, van Buuren S, van Wieringen JC, Reerink JD, Verloove-Vanhorick SP \& Ruys JH (1994). Growth in length and weight from birth to 2 years of a representative sample of Netherlands children (born in 1988-89) related to socio-economic status and other background characteristics. Annals of Human Biology, 21, 449-463.

Note

This dataset is property of the Netherlands Organisation for Applied Scientific Research TNO. Distribution is not permitted. Inquiries at stef.vanbuuren@tno.nl.