Reads selected BDS data of a person

This function takes data from a json source, validates the contents against a JSON validation schema, perform checks, calculates the D-score, calculates Z-scores and transforms the data as a tibble with a person attribute.

read_bds(
  txt = NULL,
  schema = "bds_schema.json",
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)

Arguments

txt
A JSON string, URL or file

schema
A JSON string, URL or file that selects the JSON validation schema. The default corresponds to schema = bds_schema.json and loads inst/json/bds_schema.json. The older schema = bds_schema_str.json expects that numeric BDS values are specified as characters.

append_ddi
Should DDI measures be appended?

verbose
Show verbose output for centile::y2z()

...
Pass down to jsonlite::fromJSON()

Value

A tibble with 8 columns with a person attribute

See also

Author

Stef van Buuren 2021

Examples

fn <- system.file("examples", "Laura_S2.json", package = "bdsreader")
q <- read_bds(fn)
q

#> # A tibble: 16 x 8
#>      age xname yname zname zref                        x     y      z
#>    <dbl> <chr> <chr> <chr> <chr>                   <dbl> <dbl>  <dbl>
#>  1 0     age   hgt   hgt_z nl_1997_hgt_female_nl   0     48    -1.52 
#>  2 0.101 age   hgt   hgt_z nl_1997_hgt_female_nl   0.101 53.5  -0.499
#>  3 2.04  age   hgt   hgt_z nl_1997_hgt_female_nl   2.04  90     0.67 
#>  4 0     age   wgt   wgt_z nl_1997_wgt_female_nl   0      2.95 -1.06 
#>  5 0.101 age   wgt   wgt_z nl_1997_wgt_female_nl   0.101  4.18 -0.162
#>  6 2.04  age   wgt   wgt_z nl_1997_wgt_female_nl   2.04  13.9   0.989
#>  7 0.101 age   hdc   hdc_z nl_1997_hdc_female_nl   0.101 37.6   0.418
#>  8 2.04  age   hdc   hdc_z nl_1997_hdc_female_nl   2.04  50     1.22 
#>  9 0     age   bmi   bmi_z nl_1997_bmi_female_nl   0     12.8   0.259
#> 10 0.101 age   bmi   bmi_z nl_1997_bmi_female_nl   0.101 14.6   0.231
#> 11 2.04  age   bmi   bmi_z nl_1997_bmi_female_nl   2.04  17.2   0.825
#> 12 0.101 age   dsc   dsc_z nl_2014_dsc_female_40   0.101 16.4   0.668
#> 13 2.04  age   dsc   dsc_z nl_2014_dsc_female_40   2.04  70.2   2.10 
#> 14 0     hgt   wfh   wfh_z nl_1997_wfh_female_nla 48      2.95 NA    
#> 15 0.101 hgt   wfh   wfh_z nl_1997_wfh_female_nla 53.5    4.18  0.215
#> 16 2.04  hgt   wfh   wfh_z nl_1997_wfh_female_nla 90     13.9   0.82