Reads selected BDS data of a person

This function takes data from a json source, validates the contents against a JSON validation schema, perform checks, calculates the D-score, calculates Z-scores and transforms the data as a tibble with a person attribute.

read_bds(
  txt = NULL,
  auto_format = TRUE,
  format = "1.0",
  schema = NULL,
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)

Arguments

txt
A JSON string, URL or file

auto_format
Logical. Should the format be read from the data? Default is TRUE.

format
String. JSON data schema version number. There are currently three schemas supported: 1.0, 1.1 and 2.0.

schema
A file name (optionally including the path) with the JSON validation schema. The schema argument overrides format. The function extracts the version number for the basename, and overwrites the format argument by version number.

append_ddi
Should the DDI responses be appended?

verbose
Show verbose output for centile::y2z()

...
Passed down to jsonlite::fromJSON()

Value

A tibble with 8 columns with a person attribute

Details

If txt is unspecified or NULL, then the function return will have zero rows.

The format and schema arguments specify the format of the JSON input data argument txt. The default is format = "1.0" expects that the JSON input data conform to the schema specified in system.file("schemas/bds_v1.0.json", package = "bdsreader"). The alternative is format = 2.0 expects data coded according to system.file("schemas/bds_v2.0.json", package = "bdsreader"). For new users, we recommend format "2.0".

Alternatively, the format can be specified in the JSON data file with an entry named Format. For auto_format == TRUE, the data specification overrides any format and schema arguments to the read_bds() function. The schema bds_v2.0.json schema requires the Format field, so the correct format is automatically set by the data.

If you erroneously read a JSON file of format "1.0" using format "2.0" you may see an error: Error in b[b$ElementNummer == f & !is.na(b$ElementNummer), "Waarde"] : incorrect number of dimensions. In that make sure that you are reading with the format = "1.0" argument. Reversely, if you erroneously read a JSON file of format "2.0" using format "1.0" you may see messages .ClientGegevens should be object and Missing 'ClientGegevens$Groepen'. In that case, specify format = "2.0".

See also

Author

Stef van Buuren 2021

Examples

# Assume that jamesdemodata is installed locally.
# If not use remotes::install_github("growthcharts/jamesdemodata")

# Read file with input data according to format "2.0".
data2 <- system.file("extdata/bds_v2.0/smocc/Laura_S.json", package = "jamesdemodata")
q <- read_bds(data2, format = "2.0")
q
#> # A tibble: 58 × 8
#>      age xname yname zname zref                      x     y      z
#>    <dbl> <chr> <chr> <chr> <chr>                 <dbl> <dbl>  <dbl>
#>  1 0     age   hgt   hgt_z nl_1997_hgt_female_nl 0      48   -1.52 
#>  2 0.101 age   hgt   hgt_z nl_1997_hgt_female_nl 0.101  53.5 -0.499
#>  3 0.159 age   hgt   hgt_z nl_1997_hgt_female_nl 0.159  56   -0.261
#>  4 0.236 age   hgt   hgt_z nl_1997_hgt_female_nl 0.236  59.5  0.163
#>  5 0.485 age   hgt   hgt_z nl_1997_hgt_female_nl 0.485  65.5 -0.259
#>  6 0.753 age   hgt   hgt_z nl_1997_hgt_female_nl 0.753  71.5  0.131
#>  7 1.02  age   hgt   hgt_z nl_1997_hgt_female_nl 1.02   75   -0.18 
#>  8 1.25  age   hgt   hgt_z nl_1997_hgt_female_nl 1.25   80    0.421
#>  9 1.54  age   hgt   hgt_z nl_1997_hgt_female_nl 1.54   84    0.527
#> 10 2.04  age   hgt   hgt_z nl_1997_hgt_female_nl 2.04   90    0.67 
#> # … with 48 more rows

# Equivalent, but specifying the built-in schema file bds_v2.0.json
schema2 <- system.file("schemas/bds_v2.0.json", package = "bdsreader")
r <- read_bds(data2, schema = schema2)
identical(q, r)
#> [1] TRUE

# Automatic detection of format 2.0
# s <- read_bds(data2)
# identical(q, s)

# Reading data with older format (bds_v1.0)
data1 <- system.file("extdata/bds_v1.0/smocc/Laura_S.json", package = "jamesdemodata")
t <- read_bds(data1)
t
#> # A tibble: 58 × 8
#>      age xname yname zname zref                      x     y      z
#>    <dbl> <chr> <chr> <chr> <chr>                 <dbl> <dbl>  <dbl>
#>  1 0     age   hgt   hgt_z nl_1997_hgt_female_nl 0      48   -1.52 
#>  2 0.101 age   hgt   hgt_z nl_1997_hgt_female_nl 0.101  53.5 -0.499
#>  3 0.159 age   hgt   hgt_z nl_1997_hgt_female_nl 0.159  56   -0.261
#>  4 0.236 age   hgt   hgt_z nl_1997_hgt_female_nl 0.236  59.5  0.163
#>  5 0.485 age   hgt   hgt_z nl_1997_hgt_female_nl 0.485  65.5 -0.259
#>  6 0.753 age   hgt   hgt_z nl_1997_hgt_female_nl 0.753  71.5  0.131
#>  7 1.02  age   hgt   hgt_z nl_1997_hgt_female_nl 1.02   75   -0.18 
#>  8 1.25  age   hgt   hgt_z nl_1997_hgt_female_nl 1.25   80    0.421
#>  9 1.54  age   hgt   hgt_z nl_1997_hgt_female_nl 1.54   84    0.527
#> 10 2.04  age   hgt   hgt_z nl_1997_hgt_female_nl 2.04   90    0.67 
#> # … with 48 more rows

# same, but using a built-in schema file
schema1 <- system.file("schemas/bds_v1.0.json", package = "bdsreader")
u <- read_bds(data1, schema = schema1)
identical(t, u)
#> [1] TRUE